simple_search_agent/requirements.txt
langchain
streamlit
requests
python-dotenv
openai
serpapi
langchain-openai